fixed most comments

+++ Tevling/Components/ChallengeCard.razor.cs
@@ -4,74 +4,64 @@ namespace Tevling.Components;
 
 public partial class ChallengeCard : ComponentBase
 {
-    [Inject]
-    IChallengeService ChallengeService { get; set; } = null!;
-    [Inject]
-    IJSRuntime JSRuntime { get; set; } = null!;
+    [Inject] private IChallengeService ChallengeService { get; set; } = null!;
+    [Inject] private IJSRuntime JSRuntime { get; set; } = null!;
+    [Inject] private IBrowserTime BrowserTime { get; set; } = null!;
 
-    [Parameter]
-    public int AthleteId { get; set; }
+    [Parameter] public int AthleteId { get; set; }
+    [Parameter] public Challenge? Challenge { get; set; }
 
-    [Parameter]
-    public Challenge? Challenge { get; set; }
-
+    private DateTimeOffset? CurrentBrowserTime { get; set; }
     private ScoreBoard? ScoreBoard { get; set; }
 
-    private bool ShowScoreBoard { get; set; }
-
-    private bool? HasJoined =>
-        Challenge?.Athletes?.Any(a => a.Id == AthleteId);
-
-    private Athlete? Winner {get; set; }
-    private bool DrawingWinner { get; set;}
-
-
-    private string GetIconForMeasurement()
+    private string MeasurementIcon
     {
-        return Challenge?.Measurement switch
+        get
         {
-            ChallengeMeasurement.Distance => "bi-arrow-right", 
-            ChallengeMeasurement.Time => "bi-stopwatch-fill", 
-            ChallengeMeasurement.Elevation => "bi-arrow-up", 
-            _ => "bi-question-circle", 
-        };
+            return Challenge?.Measurement switch
+            {
+                ChallengeMeasurement.Distance => "bi-arrow-right",
+                ChallengeMeasurement.Time => "bi-stopwatch-fill",
+                ChallengeMeasurement.Elevation => "bi-arrow-up",
+                _ => "bi-question-circle"
+            };
+        }
+    }
+
+    private bool DrawingWinner { get; set; }
+    private bool? HasJoined => Challenge?.Athletes?.Any(a => a.Id == AthleteId);
+    private bool ShowScoreBoard { get; set; }
+    private Athlete? Winner { get; set; }
+
+    protected override async Task OnParametersSetAsync()
+    {
+        if (Challenge != null) CurrentBrowserTime = await BrowserTime.ConvertToLocal(DateTimeOffset.Now);
     }
 
     private async Task ToggleScoreBoard()
     {
         if (Challenge != null && (ShowScoreBoard = !ShowScoreBoard))
-        {
             ScoreBoard = await ChallengeService.GetScoreBoardAsync(Challenge.Id);
-        }
     }
 
     private async Task JoinChallenge()
     {
-        if (Challenge != null)
-        {
-            Challenge = await ChallengeService.JoinChallengeAsync(AthleteId, Challenge.Id);
-        }
+        if (Challenge != null) Challenge = await ChallengeService.JoinChallengeAsync(AthleteId, Challenge.Id);
     }
 
     private async Task LeaveChallenge()
     {
-        if (Challenge != null)
-        {
-            Challenge = await ChallengeService.LeaveChallengeAsync(AthleteId, Challenge.Id);
-        }
+        if (Challenge != null) Challenge = await ChallengeService.LeaveChallengeAsync(AthleteId, Challenge.Id);
     }
 
     private async Task DeleteChallenge()
     {
-        if (Challenge != null)
-        {
-            await ChallengeService.DeleteChallengeAsync(Challenge.Id);
-        }
+        if (Challenge != null) await ChallengeService.DeleteChallengeAsync(Challenge.Id);
     }
 
-    private async Task DrawWinner() 
+    private async Task DrawWinner()
     {
-        if (Challenge is not null) 
+        if (Challenge is not null)
         {
             DrawingWinner = true;
             await Task.Delay(TimeSpan.FromSeconds(3));
